Add data modules. Add datasets. Add sampler. Add split util script. Add embedder + factory pattern. Add lstm + rnn + base model. Add train script.

+++ src/data/sampler.py
@@ -9,7 +9,7 @@ def get_weighted_sampler(labels: pd.Series) -> WeightedRandomSampler:
     counts = np.array([len(np.where(encoded_labels == t)[0]) for t in np.unique(encoded_labels)])
 
     weights = 1. / counts
-    sample_weights = torch.from_numpy(np.array([weights[t] for t in encoded_labels])).double()
+    sample_weights = torch.from_numpy(np.array([weights[t] for t in encoded_labels])).float()
 
     sampler = WeightedRandomSampler(sample_weights, len(sample_weights))
     return sampler
